Cell size is now a UInt...

+++ Game Of Life - App/Classes/Grid.swift
@@ -342,7 +342,7 @@ class Grid: NSObject
             }
         }
         
-        Preferences.shared.cellSize = CGFloat( size )
+        Preferences.shared.cellSize = UInt( size )
         
         self.population = n
         self.cells      = cells
